style(helper): :truck: rename `redis` to `RedisHelper` and export as default
