+++ src/config/routes/books/chapters/show.ts
@@ -6,7 +6,7 @@ import { RequestGeneric } from '#config/typing/requests'
 import ChapterHelper from '#helpers/books/audible/ChapterHelper'
 import addTimestamps from '#helpers/database/addTimestamps'
 import { PaprAudibleChapterHelper } from '#helpers/database/audible'
-import { RedisHelper } from '#helpers/database/redis'
+import RedisHelper from '#helpers/database/RedisHelper'
 import SharedHelper from '#helpers/shared'
 
 async function _show(fastify: FastifyInstance) {
